+++ test/test_embedded_objects.cpp
@@ -18,7 +18,7 @@ TEST(EmbeddedObjects_Basic)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-        TableRef sub = tr.add_embedded_table("class_Sub");
+        TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
         top->add_column(*sub, "sub");
         sub->add_column(type_Int, "i");
     });
@@ -49,8 +49,8 @@ TEST(Table_EmbeddedObjectsCircular)
     client_1->create_schema([&](WriteTransaction& tr) {
         Group& g = tr.get_group();
         auto table = g.add_table_with_primary_key("class_table", type_Int, "id");
-        auto e1 = g.add_embedded_table("class_e1");
-        auto e2 = g.add_embedded_table("class_e2");
+        auto e1 = g.add_table("class_e1", Table::Type::Embedded);
+        auto e2 = g.add_table("class_e2", Table::Type::Embedded);
         table->add_column(*table, "unused");
         col_link1 = table->add_column(*e1, "link");
         col_link2 = e1->add_column(*e2, "link");
@@ -78,7 +78,7 @@ TEST(EmbeddedObjects_ArrayOfObjects)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-        TableRef sub = tr.add_embedded_table("class_Sub");
+        TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
         top->add_column_list(*sub, "sub");
         sub->add_column(type_Int, "i");
     });
@@ -112,7 +112,7 @@ TEST(EmbeddedObjects_DictionaryOfObjects)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-        TableRef sub = tr.add_embedded_table("class_Sub");
+        TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
         top->add_column_dictionary(*sub, "sub");
         sub->add_column(type_Int, "i");
     });
@@ -146,7 +146,7 @@ TEST(EmbeddedObjects_NestedArray)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef threads = tr.get_group().add_table_with_primary_key("class_ForumThread", type_Int, "pk");
-        TableRef comments = tr.add_embedded_table("class_Comment");
+        TableRef comments = tr.add_table("class_Comment", Table::Type::Embedded);
         threads->add_column_list(*comments, "comments");
         comments->add_column_list(*comments, "replies");
         comments->add_column(type_Int, "message");
@@ -206,7 +206,7 @@ TEST(EmbeddedObjects_ImplicitErase)
 
         client_1->create_schema([](WriteTransaction& tr) {
             TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-            TableRef sub = tr.add_embedded_table("class_Sub");
+            TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
             top->add_column(*sub, "sub");
             sub->add_column(type_Int, "i");
         });
@@ -256,7 +256,7 @@ TEST(EmbeddedObjects_SetDefaultNullIgnored)
 
         client_1->create_schema([](WriteTransaction& tr) {
             TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-            TableRef sub = tr.add_embedded_table("class_Sub");
+            TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
             top->add_column(*sub, "sub");
             sub->add_column(type_Int, "i");
         });
@@ -307,7 +307,7 @@ TEST(EmbeddedObjects_DiscardThroughImplicitErase)
 
         client_1->create_schema([](WriteTransaction& tr) {
             TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-            TableRef sub = tr.add_embedded_table("class_Sub");
+            TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
             top->add_column(*sub, "sub");
             sub->add_column(type_Int, "i");
 
@@ -356,7 +356,7 @@ TEST(EmbeddedObjects_AdjustPathOnInsert)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-        TableRef sub = tr.add_embedded_table("class_Sub");
+        TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
         top->add_column_list(*sub, "sub");
         sub->add_column_list(*sub, "sub");
         sub->add_column(type_Int, "i");
@@ -430,7 +430,7 @@ TEST(EmbeddedObjects_AdjustPathOnErase)
 
     client_1->create_schema([](WriteTransaction& tr) {
         TableRef top = tr.get_group().add_table_with_primary_key("class_Top", type_Int, "pk");
-        TableRef sub = tr.add_embedded_table("class_Sub");
+        TableRef sub = tr.add_table("class_Sub", Table::Type::Embedded);
         top->add_column_list(*sub, "sub");
         sub->add_column_list(*sub, "sub");
         sub->add_column(type_Int, "i");
@@ -508,7 +508,7 @@ TEST(EmbeddedObjects_CreateEraseCreateSequencePreservesObject)
         client_1->transaction([&](Peer& c) {
             auto& tr = *c.group;
             auto table = tr.add_table_with_primary_key("class_table", type_Int, "pk");
-            auto embedded = tr.add_embedded_table("class_embedded");
+            auto embedded = tr.add_table("class_embedded", Table::Type::Embedded);
             embedded->add_column(type_Int, "int");
             table->add_column(*embedded, "embedded");
             auto obj = table->create_object_with_primary_key(123);
@@ -575,7 +575,7 @@ TEST(EmbeddedObjects_CreateEraseCreateSequencePreservesObject_Nested)
         client_1->transaction([&](Peer& c) {
             auto& tr = *c.group;
             auto table = tr.add_table_with_primary_key("class_table", type_Int, "pk");
-            auto embedded = tr.add_embedded_table("class_embedded");
+            auto embedded = tr.add_table("class_embedded", Table::Type::Embedded);
             embedded->add_column(type_Int, "int");
             embedded->add_column(*embedded, "embedded");
             table->add_column(*embedded, "embedded");